Se reestructura el controlador para editar empleados
