+++ server/controllers/usuarioController/Empleador/editarEmpleadorController.js
@@ -1,47 +1,17 @@
-const UsuarioModel = require('../../../models/usuarioModel/usuariosModel.js');
+const empleadorServicios = require('../../../servicios/usuarioServicios/empleadorServicios/empleadorServicios');
+const usuarioModelo = require('../../../models/usuarioModel/usuariosModel');
 
-async function editarPerfilEmpleador(solicitud, respuesta) {
-    try {
-        const datos = solicitud.body;
-
-        let telefono2 = null;
-
-        if (datos["telefono2"] != "") {
-            telefono2 = datos["telefono2"];
-        }
-
-        let correo2 = null;
-        if (datos["correo2"] != "") {
-            correo2 = datos["correo2"];
-        }
-
-        var exito = await UsuarioModel.editarPerfilEmpleador(
-            datos["cedula"],
-            datos["nombre"],
-            datos["primer_apellido"],
-            datos["segundo_apellido"],
-            datos["telefono1"],
-            telefono2,
-            datos["correo1"],
-            correo2
-        );
-
-        
-        return exito;
-    }
-    catch (error) {
-        return error;
-    }
-}
-
+//Funcion que verifica si existe un usuario y de haberlo procesa la edicion
 async function EditarEmpleador(solicitud, respuesta) {
     try {
         const datos = solicitud.body;
         const cedula = datos["cedula"];
-        const existeUsuario = await UsuarioModel.getByCedula(cedula);
+        //Funcion que verifica si existe el usuario
+        const existeUsuario = await usuarioModelo.getByCedula(cedula);
 
         if (existeUsuario) {
-            const exito = await editarPerfilEmpleador(solicitud, respuesta);
+            //Se procesa la edicion del empleador a traves de la funcion del servicio
+            const exito = await empleadorServicios.editarPerfilEmpleador(solicitud, respuesta);
             if (exito) { 
                 respuesta.status(201).json({ message: 'Empleado editado exitosamente' });
             } else {
